AI模型/simple_model/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
joblib==1.0.1
requests==2.26.0
psycopg2-binary==2.9.1
gunicorn==20.1.0